requirements.txt
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
dotenv==0.9.9
executing==2.2.1
fastjsonschema==2.21.2
Flask==3.1.2
greenlet==3.2.4
idna==3.11
iniconfig==2.3.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyzmq==27.1.0
redis==6.4.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.5.3
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.2
wcwidth==0.2.14
webencodings==0.5.1
Werkzeug==3.1.3
yarg==0.1.9
